LPS-55590 Copy EditOrganizationAssignmentsAction, no logic changes
